feat: all images validated equally
