+++ framework/core/src/Admin/FaviconValidator.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Flarum\Admin;
+
+use Flarum\Foundation\AbstractImageValidator;
+
+class FaviconValidator extends AbstractImageValidator
+{
+
+}
