+++ framework/core/src/Admin/LogoValidator.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Flarum\Admin;
+
+use Flarum\Foundation\AbstractImageValidator;
+
+class LogoValidator extends AbstractImageValidator
+{
+
+}
